refactor: KGQueryStore leaves the view layer (observer-pattern + view-endpoint gates)
OntologyBrowser+Sheets.swift called client.api.* from a VIEW file and scraped its
client off LibraryManager.shared.globalLibrary — the two things the
observable-data-layer rule (#1863) forbids. Fixed properly, not allowlisted.

KGQueryStore was ALREADY a correct @Observable store owning the SPARQL calls; it
was just declared inside a view file and hand-fed a singleton-scraped client. So:

- Moved KGQueryStore verbatim → Models/KGQueryStore.swift. Endpoint access now
  lives in a store file, not under Views/. (+ pbxproj membership: the app target
  is not file-synchronized, only the test targets are.)
- AppState owns it (app-wide, like sessionStore/usersStore/identityStore) and
  builds it from the app-wide FicheroClient.
- Injected at the canonical DI seam in DocumentTabView, next to entityStore /
  claimStore — the same seam #3033 established.
- KGQueryConsoleView drops `let client` + its custom init and reads
  @Environment(KGQueryStore.self). The sheet is now just KGQueryConsoleView(),
  with no LibraryManager.shared anywhere in the file.

Net: the view observes a store, the store is the only endpoint accessor, and the
console now shares one store instance with the rest of the app instead of minting
a fresh one per sheet presentation.

check_observer_pattern → 0, check_view_endpoint_access → 0. Both scripts also
report pre-existing STALE allowlist entries (7 and 6) — left alone, not mine to
churn in this commit. swiftlint clean. Commit-only.

(cherry picked from commit 2029e5e54163635932462db70d561a5dda7f118f)
